[STABLE] Work queue idea.
